Location reminder service fully functional, works on reboot

+++ RemindMeWhenIAmThere/app/src/main/java/com/vinkel/remindmewheniamthere/background/observers/base/ILocationObserver.java
@@ -4,7 +4,7 @@ import android.location.Location;
 import io.reactivex.Observable;
 
 public interface ILocationObserver {
-  Observable<Location> getLocationObserver(final long minUpdateTime, final long minUpdateDistance, final String observerTag);
+  Observable<Location> getLocationObserver(final long minUpdateTime, final float minUpdateDistance, final String observerTag);
 
   Observable<Location> getLocationOnce();
 }
